requirements.txt
asgiref==3.8.1
Django==4.1.13
django-filter==23.5
djangorestframework==3.15.1
djongo==1.3.6
dnspython==2.6.1
Markdown==3.6
mysqlclient==2.2.4
pymongo==4.6.3
sqlparse==0.2.4
